:truck: Rename FileUploadView to FilesView
