+++ src/views/index.js
@@ -1,3 +1,3 @@
 export {default as StudyListView} from './StudyListView';
-export {default as FileUploadView} from './FileUploadView';
+export {default as FilesView} from './FilesView';
 export {default as LoginView} from './LoginView';
